pyproject.toml
[project]
name = "interview-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.3.0",
    "google-genai>=1.0.0",
    "kokoro>=0.7.13",
    "pip>=25.0.1",
    "python-dotenv>=1.0.1",
    "realtimestt>=0.3.94",
    "soundfile>=0.13.1",
    "transformers>=4.48.3",
]
